t7v2

+++ 0x00-ES6_basic/7-getBudgetObject.js
@@ -1,8 +1,8 @@
 export default function getBudgetObject(income, gdp, capita) {
   const budget = {
-    income: int = income,
-    gdp: int = gdp,
-    capita: int = capita,
+    income: float = income,
+    gdp: float = gdp,
+    capita: float = capita,
   };
 
   return budget;
